+++ CHANGES.md
@@ -4,6 +4,9 @@ Unreleased
 - hpack: fix a case where hpack would raise an array out of bounds exception
   ([#183](https://github.com/anmonteiro/ocaml-h2/pull/183))
   ([@jonathanjameswatson](https://github.com/jonathanjameswatson))
+- h2: (client) handle multiple RST_STREAM frames
+  ([#184](https://github.com/anmonteiro/ocaml-h2/pull/184))
+  ([@jonathanjameswatson](https://github.com/jonathanjameswatson))
 
 0.9.0 2022-08-14
 ---------------